decimal weight, price

+++ AdminApp/src/dao/ProductsHandler.java
@@ -1,5 +1,6 @@
 package dao;
 
+import java.math.BigDecimal;
 import java.sql.Connection;
 import java.sql.PreparedStatement;
 import java.sql.ResultSet;
@@ -23,8 +24,8 @@ public class ProductsHandler {
 				 p.setProductDesc(rs.getString("productDesc"));
 				 p.setProductId(rs.getInt("productId"));
 				 p.setProductName(rs.getString("productName"));
-				 p.setWeight_lb(rs.getFloat("weight_lb"));
-				 p.setPrice(rs.getFloat("price"));
+				 p.setWeight_lb(rs.getBigDecimal("weight_lb"));
+				 p.setPrice(rs.getBigDecimal("price"));
 				 productsList.add(p);
 			 }
 			 //System.out.println(productsList);
@@ -47,8 +48,8 @@ public class ProductsHandler {
 				String name = p.getProductName();
 				String cat = p.getProductCategory();
 				String desc = p.getProductDesc();
-				float weight = p.getWeight_lb();
-				float price = p.getPrice();
+				BigDecimal weight = p.getWeight_lb();
+				BigDecimal price = p.getPrice();
 								
 				String q1 = "SELECT MAX(productId) AS max FROM products";
 				PreparedStatement ps1 = c.prepareStatement(q1);
@@ -87,8 +88,8 @@ public class ProductsHandler {
 					 p.setProductDesc(rs.getString("productDesc"));
 					 p.setProductId(rs.getInt("productId"));
 					 p.setProductName(rs.getString("productName"));
-					 p.setWeight_lb(rs.getFloat("weight_lb"));
-					 p.setPrice(rs.getFloat("price"));
+					 p.setWeight_lb(rs.getBigDecimal("weight_lb"));
+					 p.setPrice(rs.getBigDecimal("price"));
 					 
 				 }
 				 //System.out.println(productsList);
@@ -112,8 +113,8 @@ public class ProductsHandler {
 				String name = p.getProductName();
 				String cat = p.getProductCategory();
 				String desc = p.getProductDesc();
-				float weight = p.getWeight_lb();
-				float price = p.getPrice();
+				BigDecimal weight = p.getWeight_lb();
+				BigDecimal price = p.getPrice();
 								
 		                
 		         String q = "UPDATE products SET productName='"+name+"', productCategory='"+cat+"', productDesc='"+desc+"', weight_lb="+weight+", price="+price+" WHERE productId = "+pid;
